fix u-boot packages

+++ oe/recipes/u-boot/u-boot-igep_2010.06-3.bb
@@ -16,39 +16,61 @@ SRC_URI = "http://downloads.igep.es/sources/${PN}-${PV}.tar.gz \
 	file://fw_env.config "
 
 PACKAGE_ARCH = "${MACHINE_ARCH}"
-
 PARALLEL_MAKE = ""
 
+EXTRA_OEMAKE = "CROSS_COMPILE=${TARGET_PREFIX}"
+
+UBOOT_MACHINE = "igep0030_sd_config"
+
+UBOOT_BINARY ?= "u-boot.bin"
+UBOOT_IMAGE ?= "u-boot-${MACHINE}-${PV}-${PR}.bin"
+UBOOT_SYMLINK ?= "u-boot-${MACHINE}.bin"
+UBOOT_MAKE_TARGET ?= "all"
+
+do_configure () {
+	oe_runmake ${UBOOT_MACHINE}
+}
+
 do_compile () {
-	unset LDFLAGS CFLAGS CPPFLAGS
-	oe_runmake ${MACHINE}_config
-	oe_runmake CROSS_COMPILE=${TARGET_PREFIX}
-	oe_runmake CROSS_COMPILE=${TARGET_PREFIX} env tools
+	unset LDFLAGS
+	unset CFLAGS
+	unset CPPFLAGS
+	oe_runmake ${UBOOT_MAKE_TARGET}
+	oe_runmake tools env HOSTCC="${CC}"
 }
 
 do_install () {
-	install -d ${D}${bindir}
-	install -m 0755 ${S}/tools/env/fw_printenv ${D}${bindir}
-	ln -sf fw_printenv ${D}${bindir}/fw_setenv
-	install -m 0744 ${S}/tools/mkimage ${D}${bindir}
+	install -d ${D}/boot
+	install ${S}/${UBOOT_BINARY} ${D}/boot/${UBOOT_IMAGE}
+	ln -sf ${UBOOT_IMAGE} ${D}/boot/${UBOOT_BINARY}
 
-	install -d ${D}${sysconfdir}
-	install -m 0666 ${WORKDIR}/fw_env.config ${D}${sysconfdir}
+	if [ -e ${WORKDIR}/fw_env.config ] ; then
+		install -d ${D}${base_sbindir}
+		install -d ${D}${sysconfdir}
+		install -m 644 ${WORKDIR}/fw_env.config ${D}${sysconfdir}/fw_env.config
+		install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_printenv
+	install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_setenv
+	fi
 }
+
+FILES_${PN} = "/boot"
+# no gnu_hash in uboot.bin, by design, so skip QA
+INSANE_SKIP_${PN} = True
+
+PACKAGES += "${PN}-fw-utils"
+FILES_${PN}-fw-utils = "${sysconfdir} ${base_sbindir}"
+# u-boot doesn't use LDFLAGS for fw files, needs to get fixed, but until then:
+INSANE_SKIP_${PN}-fw-utils = True
 
 do_deploy () {
 	install -d ${DEPLOY_DIR_IMAGE}
-	install -m 0666 ${S}/u-boot.bin ${DEPLOY_DIR_IMAGE}/u-boot-${PV}.${MACHINE}.bin
-	package_stagefile_shell ${DEPLOY_DIR_IMAGE}/u-boot-${PV}.${MACHINE}.bin
+	install ${S}/${UBOOT_BINARY} ${DEPLOY_DIR_IMAGE}/${UBOOT_IMAGE}
+	package_stagefile_shell ${DEPLOY_DIR_IMAGE}/${UBOOT_IMAGE}
 
-	rm -f u-boot-${MACHINE}.bin
-	ln -sf u-boot-${PV}.${MACHINE}.bin ${DEPLOY_DIR_IMAGE}/u-boot-${MACHINE}.bin
-	package_stagefile_shell ${DEPLOY_DIR_IMAGE}/u-boot-${MACHINE}.bin
+	cd ${DEPLOY_DIR_IMAGE}
+	rm -f ${UBOOT_SYMLINK}
+	ln -sf ${UBOOT_IMAGE} ${UBOOT_SYMLINK}
+	package_stagefile_shell ${DEPLOY_DIR_IMAGE}/${UBOOT_SYMLINK}
 }
 do_deploy[dirs] = "${S}"
-addtask deploy before do_build after do_compile
-
-PACKAGES = "u-boot-env u-boot-env-dbg"
-
-FILES_u-boot-env = "${bindir}/fw_* ${sysconfdir} ${bindir}/mkimage"
-FILES_u-boot-env-dbg = "${bindir}/.debug"
+addtask deploy before do_package_stage after do_compile
